+++ packages/executor-react/src/__tests__/midlConfig.ts
@@ -1,13 +1,12 @@
 import { createConfig, regtest } from "@midl-xyz/midl-js-core";
 import { keyPairConnector } from "@midl-xyz/midl-js-node";
-import * as bitcoin from "bitcoinjs-lib";
-import { getKeyPair } from "~/__tests__/keyPair";
+import { __TEST__MNEMONIC__ } from "~/__tests__/keyPair";
 
 export const midlConfig = createConfig({
 	networks: [regtest],
 	connectors: [
 		keyPairConnector({
-			keyPair: getKeyPair(bitcoin.networks.regtest),
+			mnemonic: __TEST__MNEMONIC__,
 		}),
 	],
 });