+++ packages/satoshi-kit/src/__tests__/wrapper.tsx
@@ -3,7 +3,7 @@ import { keyPairConnector } from "@midl-xyz/midl-js-node";
 import { MidlProvider } from "@midl-xyz/midl-js-react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import type { ReactNode } from "react";
-import { getKeyPair } from "~/__tests__/keyPair";
+import { __TEST__MNEMONIC__ } from "~/__tests__/keyPair";
 import { SatoshiKitProvider } from "~/app";
 import type { AuthenticationAdapter } from "~/features/auth";
 
@@ -15,10 +15,10 @@ const queryClient = new QueryClient({
 	},
 });
 
-export const testConnector = keyPairConnector({ keyPair: getKeyPair() });
+export const testConnector = keyPairConnector({ mnemonic: __TEST__MNEMONIC__ });
 
 export const midlConfig = createConfig({
-	connectors: [keyPairConnector({ keyPair: getKeyPair() })],
+	connectors: [keyPairConnector({ mnemonic: __TEST__MNEMONIC__ })],
 	networks: [regtest],
 });