+++ packages/node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-node
 
+## 2.0.0-next.58
+
+### Patch Changes
+
+- feat(hardhat-deploy): segwit
+- Updated dependencies
+  - @midl-xyz/midl-js-core@2.0.0-next.58
+
 ## 2.0.0-next.54
 
 ### Patch Changes